add name for signer + auto open generateCommitmentModal when connect wallet
